refactor: extensive improvements

+++ main/main.cpp
@@ -7,11 +7,9 @@
 #include <QStandardPaths>
 #include <QLocalSocket>
 #include <QLocalServer>
-#include <QThread>
+#include <QCryptographicHash>
 
-#include "3rdparty/RunGuard.hpp"
 #include "main/NekoGui.hpp"
-
 #include "ui/mainwindow.h"
 
 #ifdef Q_OS_WIN
@@ -20,66 +18,37 @@
 
 void signal_handler(int signum) {
     if (qApp) {
-        GetMainWindow()->on_commitDataRequest();
+        MainWindow::instance()->on_commitDataRequest();
         qApp->exit();
     }
 }
 
-QTranslator* trans = nullptr;
-QTranslator* trans_qt = nullptr;
-
-void loadTranslate(const QString& locale) {
-    if (trans != nullptr) {
-        trans->deleteLater();
-    }
-    if (trans_qt != nullptr) {
-        trans_qt->deleteLater();
-    }
-    //
-    trans = new QTranslator;
-    trans_qt = new QTranslator;
-    QLocale::setDefault(QLocale(locale));
-    //
-    if (trans->load(":/translations/" + locale + ".qm")) {
-        QCoreApplication::installTranslator(trans);
-    }
-    if (trans_qt->load(QApplication::applicationDirPath() + "/qtbase_" + locale + ".qm")) {
-        QCoreApplication::installTranslator(trans_qt);
-    }
-}
-
-#define LOCAL_SERVER_PREFIX "nekoraylocalserver-"
-
-int main(int argc, char* argv[]) {
+int main(int argc, char *argv[]) {
     // Core dump
 #ifdef Q_OS_WIN
     Windows_SetCrashHandler();
 #endif
 
     // pre-init QApplication
-#if QT_VERSION < QT_VERSION_CHECK(6, 0, 0) && QT_VERSION >= QT_VERSION_CHECK(5, 10, 0)
-    QApplication::setAttribute(Qt::AA_DisableWindowContextHelpButton);
-#endif
-#if QT_VERSION >= QT_VERSION_CHECK(5, 7, 0)
-    QApplication::setAttribute(Qt::AA_DontUseNativeDialogs);
-#endif
-    QApplication::setQuitOnLastWindowClosed(false);
-    auto preQApp = new QApplication(argc, argv);
+    QApplication app(argc, argv);
 
-    // Clean
-    QDir::setCurrent(QApplication::applicationDirPath());
-    if (QFile::exists("updater.old")) {
-        QFile::remove("updater.old");
-    }
-#ifndef Q_OS_WIN
-    if (!QFile::exists("updater")) {
-        QFile::link("launcher", "updater");
-    }
-#endif
+    // RunGuard
+    const QString APP_KEY = QCryptographicHash::hash(QApplication::applicationFilePath().toUtf8(), QCryptographicHash::Md5).toHex();
+    QLocalSocket socket;
+    socket.connectToServer(APP_KEY);
+    if (socket.waitForConnected(100)) return 0;
+    QLocalServer::removeServer(APP_KEY);
+    QLocalServer server;
+    server.listen(APP_KEY);
+    QObject::connect(&server, &QLocalServer::newConnection, &app, [&] {
+        auto socket = server.nextPendingConnection();
+        socket->deleteLater();
+        // raise main window
+        MW_dialog_message("", "Raise");
+    });
 
     // Flags
     NekoGui::dataStore->argv = QApplication::arguments();
-    if (NekoGui::dataStore->argv.contains("-many")) NekoGui::dataStore->flag_many = true;
     if (NekoGui::dataStore->argv.contains("-appdata")) {
         NekoGui::dataStore->flag_use_appdata = true;
         int appdataIndex = NekoGui::dataStore->argv.indexOf("-appdata");
@@ -99,88 +68,32 @@ int main(int argc, char* argv[]) {
 #endif
 
     // dirs & clean
-    auto wd = QDir(QApplication::applicationDirPath());
+    QDir appDir(QApplication::applicationDirPath());
     if (NekoGui::dataStore->flag_use_appdata) {
         QApplication::setApplicationName("nekoray");
         if (!NekoGui::dataStore->appdataDir.isEmpty()) {
-            wd.setPath(NekoGui::dataStore->appdataDir);
+            appDir.setPath(NekoGui::dataStore->appdataDir);
         } else {
-            wd.setPath(QStandardPaths::writableLocation(QStandardPaths::AppConfigLocation));
+            appDir.setPath(QStandardPaths::writableLocation(QStandardPaths::AppConfigLocation));
         }
     }
-    if (!wd.exists()) wd.mkpath(wd.absolutePath());
-    if (!wd.exists("config")) wd.mkdir("config");
-    QDir::setCurrent(wd.absoluteFilePath("config"));
+    if (!appDir.mkpath("config")) {
+        QMessageBox::critical(nullptr, "Error", "Cannot create config directory:\n" + appDir.absoluteFilePath("config"));
+        return 1;
+    }
+    QDir::setCurrent(appDir.absoluteFilePath("config"));
     QDir("temp").removeRecursively();
-
-    // init QApplication
-    delete preQApp;
-    QApplication a(argc, argv);
+    for (const auto &dir: {"profiles", "groups", "routes"}) {
+        QDir().mkdir(dir);
+    }
 
     // dispatchers
     DS_cores = new QThread;
     DS_cores->start();
 
-    // RunGuard
-    RunGuard guard("nekoray" + wd.absolutePath());
-    quint64 guard_data_in = GetRandomUint64();
-    quint64 guard_data_out = 0;
-    if (!NekoGui::dataStore->flag_many && !guard.tryToRun(&guard_data_in)) {
-        // Some Good System
-        if (guard.isAnotherRunning(&guard_data_out)) {
-            // Wake up a running instance
-            QLocalSocket socket;
-            socket.connectToServer(LOCAL_SERVER_PREFIX + Int2String(guard_data_out));
-            qDebug() << socket.fullServerName();
-            if (!socket.waitForConnected(500)) {
-                qDebug() << "Failed to wake a running instance.";
-                return 0;
-            }
-            qDebug() << "connected to local server, try to raise another program";
-            return 0;
-        }
-        // Some Bad System
-        QMessageBox::warning(nullptr, "NekoRay", "RunGuard disallow to run, use -many to force start.");
-        return 0;
-    }
-    MF_release_runguard = [&] { guard.release(); };
-
-// icons
-#if (QT_VERSION >= QT_VERSION_CHECK(5, 11, 0))
-    QIcon::setFallbackSearchPaths(QStringList{
-        ":/nekoray",
-        ":/icon",
-    });
-#endif
-
-    // icon for no theme
-    if (QIcon::themeName().isEmpty()) {
-        QIcon::setThemeName("breeze");
-    }
-
-    // Dir
-    QDir dir;
-    bool dir_success = true;
-    if (!dir.exists("profiles")) {
-        dir_success &= dir.mkdir("profiles");
-    }
-    if (!dir.exists("groups")) {
-        dir_success &= dir.mkdir("groups");
-    }
-    if (!dir.exists("routes")) {
-        dir_success &= dir.mkdir("routes");
-    }
-    if (!dir_success) {
-        QMessageBox::warning(nullptr, "Error", "No permission to write " + dir.absolutePath());
-        return 1;
-    }
-
     // Load dataStore
     NekoGui::dataStore->fn = "nekobox.json";
-    auto isLoaded = NekoGui::dataStore->Load();
-    if (!isLoaded) {
-        NekoGui::dataStore->Save();
-    }
+    if (!NekoGui::dataStore->Load()) NekoGui::dataStore->Save();
 
     // Datastore & Flags
     if (NekoGui::dataStore->start_minimal) NekoGui::dataStore->flag_tray = true;
@@ -188,48 +101,25 @@ int main(int argc, char* argv[]) {
     // load routing
     NekoGui::dataStore->routing = std::make_unique<NekoGui::Routing>();
     NekoGui::dataStore->routing->fn = "routes/" + NekoGui::dataStore->active_routing;
-    isLoaded = NekoGui::dataStore->routing->Load();
-    if (!isLoaded) {
-        NekoGui::dataStore->routing->Save();
-    }
+    if (!NekoGui::dataStore->routing->Load()) NekoGui::dataStore->routing->Save();
 
     // Translate
-    QString locale;
-    switch (NekoGui::dataStore->language) {
-        case 1: // English
-            break;
-        case 2:
-            locale = "zh_CN";
-            break;
-        case 3:
-            locale = "fa_IR"; // farsi(iran)
-            break;
-        case 4:
-            locale = "ru_RU"; // Russian
-            break;
-        default:
-            locale = QLocale().name();
+    if (NekoGui::dataStore->language.isEmpty()) NekoGui::dataStore->language = QLocale::system().name();
+    QLocale::setDefault(QLocale(NekoGui::dataStore->language));
+    QTranslator trans;
+    if (trans.load(NekoGui::dataStore->language, ":/i18n")) QApplication::installTranslator(&trans);
+
+    // Font
+    if (!NekoGui::dataStore->font.isEmpty()) {
+        QFont currentFont = QApplication::font();
+        currentFont.setFamily(NekoGui::dataStore->font);
+        QApplication::setFont(currentFont);
     }
-    QGuiApplication::tr("QT_LAYOUT_DIRECTION");
-    loadTranslate(locale);
 
     // Signals
     signal(SIGTERM, signal_handler);
     signal(SIGINT, signal_handler);
 
-    // QLocalServer
-    QLocalServer server;
-    auto server_name = LOCAL_SERVER_PREFIX + Int2String(guard_data_in);
-    QLocalServer::removeServer(server_name);
-    server.listen(server_name);
-    QObject::connect(&server, &QLocalServer::newConnection, &a, [&] {
-        auto socket = server.nextPendingConnection();
-        qDebug() << "nextPendingConnection:" << server_name << socket;
-        socket->deleteLater();
-        // raise main window
-        MW_dialog_message("", "Raise");
-    });
-
-    new MainWindow;
+    MainWindow w;
     return QApplication::exec();
 }
